// server/src/model/SubState.ts
export enum SubState {
  DAYTIME_FIRST = "First day",
  DAYTIME_VOTING = "Voting...",
  DAYTIME_TIEDVOTE = "Vote tied!",
  DAYTIME_POSTVOTE = "End of day",
  DAYTIME_GAMEFINISHED = "Game finished!",
  NIGHTTIME_SELECTION = "Night",
  NIGHTTIME_WOLVES = "Night - Wolves are hunting",
  NIGHTTIME_MEDIUM = "Night - Medium is concentrating...",
  GAME_FINISHED = "Game Finished"
}